added scrollytelling/animation for scrolling. interactive charts stay on the bottom but I didn't add and styling for those graphs.
I kept most of the previous code, so whatever charts you have that are interactive should stay below the storytelling board.

If making a visualization/animation, make sure you create a function and add it to the const charts, it has a comment above it too. There shouldn't be any need to resize the page, unless your charts are large and I can try to fix that.